pyproject.toml
[project]
name = "core-structure"
version = "0.1"
description = "Система структурного анализа и обработки документов с поддержкой удалённого OCR, иерархического управления проектами и интеллектуальной разметки блоков"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Core Structure Team"}
]
keywords = [
    "pdf",
    "ocr",
    "annotation",
    "document-processing",
    "computer-vision",
    "structure-analysis"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    "PySide6>=6.10.2,<7",
    "PyMuPDF>=1.27.2,<2",
    "Pillow>=12.1.1,<13",
    "httpx>=0.28.1,<1",
    "python-dotenv>=1.2.2,<2",
    "boto3>=1.42.68,<2",
    "PyYAML>=6.0.3,<7",
]

[project.optional-dependencies]
build = [
    "pyinstaller>=6.19.0,<7",
]

[project.urls]
Homepage = "https://github.com/your-org/core-structure"
Documentation = "https://github.com/your-org/core-structure/docs"
Repository = "https://github.com/your-org/core-structure"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app*", "rd_core*", "services*"]

[tool.setuptools.package-data]
"services.remote_ocr.server" = ["config.yaml"]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["rd_core.r2_utils", "rd_core.r2_upload", "rd_core.r2_download"]
# R2 mixins обращаются к self.s3_client/bucket_name из главного класса
disable_error_code = ["attr-defined"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
